Partially integrated with Platformer code

+++ src/InteractiveSprite.h
@@ -9,9 +9,11 @@ public:
 	InteractHandler();
 	~InteractHandler();
 
-	bool action_MouseHover(sf::Vector2f mousePos, sf::Sprite sprite);
-	bool action_MouseClick(sf::Vector2f mousePos, sf::Sprite sprite);
+	bool action_MouseHover(sf::Vector2i mousePos, sf::Sprite sprite);
+	bool action_MouseClick(sf::Vector2i mousePos, sf::Sprite sprite);
+	
 
+	
 protected:
 	
 	virtual void onHover();
@@ -21,7 +23,9 @@ protected:
 private:
 	sf::Clock inputClock;
 	float inputBuffer = INPUT_BUFFER_TIME;
-	virtual bool isMouseHovered(sf::Vector2f mousePos, sf::Sprite sprite);
-	virtual bool isMouseClicked(sf::Vector2f mousePos, sf::Sprite sprite);
+	virtual bool isMouseHovered(sf::Vector2i mousePos, sf::Sprite sprite);
+	virtual bool isMouseClicked(sf::Vector2i mousePos, sf::Sprite sprite);
+
+	
 	void resetBuffer() { inputBuffer = INPUT_BUFFER_TIME };
 };
